tests/functional/requirements.txt
pytest==7.2.2
pytest-asyncio==0.21.0
aiohttp==3.8.4

pydantic==1.10.5
motor==3.1.2
python-jose[cryptography]==3.3.0

psycopg2==2.9.6
aiopg==1.4.0